+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Added `12/30` and `12/300` problems
 
+### Changed
+- Changed `decimal().with_decimals(N)` to `decimals(N)` in `num_gen`
+
 ## [0.12.3] - 2026-08-03
 
 ### Added